Fix a 64bit problem with libpng.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@688 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/libpng/pngconf.h
@@ -1093,8 +1093,8 @@
  * want to have unsigned int for png_uint_32 instead of unsigned long.
  */
 
-typedef unsigned long png_uint_32;
-typedef long png_int_32;
+typedef unsigned int png_uint_32;
+typedef int png_int_32;
 typedef unsigned short png_uint_16;
 typedef short png_int_16;
 typedef unsigned char png_byte;
